Homework 2 updated

+++ homework-4.py
@@ -1,6 +1,8 @@
 from math import sqrt
-a = float(input('Enter first side: '))
-b = float(input('Enter second side: '))
-c = float(input('Enter third side: '))
+a = int(input('Enter first side: '))
+b = int(input('Enter second side: '))
+c = int(input('Enter third side: '))
+
 p = (a + b + c)/2
+
 print(f'Triangle area is {sqrt(p*(p-a)*(p-b)*(p-c))}')
